update file structure

+++ src/main/kotlin/com/marufh/pathagar/author/service/AuthorService.kt
@@ -1,5 +1,6 @@
 package com.marufh.pathagar.author.service
 
+import com.marufh.pathagar.author.dto.AuthorDetailsDto
 import com.marufh.pathagar.author.dto.AuthorDto
 import com.marufh.pathagar.author.dto.AuthorMapper
 import com.marufh.pathagar.author.repository.AuthorRepository
@@ -7,7 +8,10 @@ import com.marufh.pathagar.book.dto.BookMapper
 import com.marufh.pathagar.book.repository.BookRepository
 import com.marufh.pathagar.config.FileProperties
 import com.marufh.pathagar.exception.NotFoundException
-import jakarta.persistence.EntityNotFoundException
+import com.marufh.pathagar.file.dto.FileDto
+import com.marufh.pathagar.file.entity.FileType
+import com.marufh.pathagar.file.service.FileUploadService
+import com.marufh.pathagar.file.service.ImageResizeService
 import org.slf4j.LoggerFactory
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -18,6 +22,8 @@ import java.nio.file.Path
 
 @Service
 class AuthorService(
+    private val imageResizeService: ImageResizeService,
+    private val fileUploadService: FileUploadService,
     private val bookRepository: BookRepository,
     private val fileProperties: FileProperties,
     private val authorMapper: AuthorMapper,
@@ -26,11 +32,25 @@ class AuthorService(
 
     private val logger = LoggerFactory.getLogger(javaClass)
 
+    @Transactional
     fun create(authorDto: AuthorDto): AuthorDto {
         logger.info("Creating author: ${authorDto.name}")
 
-        authorMapper.toEntity(authorDto).run {
-            return authorMapper.toDto(authorRepository.save(this))
+        if(authorDto.file == null) {
+            throw IllegalArgumentException("File is required")
+        }
+
+        val authorImage = fileUploadService.createFile(FileDto(
+            name = authorDto.name,
+            file = authorDto.file!!,
+            fileType = FileType.AUTHOR
+        ))
+        val authorImageThumb = imageResizeService.createThumb(Path.of(fileProperties.base +"/"+ authorImage.path), FileType.AUTHOR)
+
+        return authorMapper.toEntity(authorDto).let {
+            it.imageFile = authorImage
+            it.thumbFile = authorImageThumb
+            authorMapper.toDto(authorRepository.save(it))
         }
     }
 
@@ -39,23 +59,34 @@ class AuthorService(
         logger.info("Updating author: ${authorDto.name}")
 
         val author = authorRepository.findById(authorDto.id!!)
-            .orElseThrow { EntityNotFoundException("Author not found with id: ${authorDto.id}") }
+            .orElseThrow { NotFoundException("Author not found with id: ${authorDto.id}") }
+
         author.name = authorDto.name
         author.description = authorDto.description
+        if (authorDto.file != null) {
+            val authorImage = fileUploadService.createFile(FileDto(
+                name = authorDto.name,
+                file = authorDto.file!!,
+                fileType = FileType.AUTHOR
+            ))
+            val authorImageThumb = imageResizeService.createThumb(Path.of(fileProperties.base +"/"+ authorImage.path), FileType.AUTHOR)
+            author.imageFile = authorImage
+            author.thumbFile = authorImageThumb
+        }
         return authorRepository.save(author).run { authorMapper.toDto(this) }
     }
 
     @Transactional
-    fun findById(id: String): AuthorDto {
+    fun findById(id: String): AuthorDetailsDto {
         logger.info("Finding author by id: $id")
 
         return authorRepository.findById(id)
-            .map {authorMapper.toDto(it) }
+            .map {authorMapper.toDetailsDto(it) }
             .orElseThrow { NotFoundException("Author not found with id: $id") }
     }
 
     @Transactional
-    fun getAuthorDetails(id: String): AuthorDto {
+    fun getAuthorDetails(id: String): AuthorDetailsDto {
         logger.info("Getting author details $id")
 
         return findById(id).apply {
@@ -73,22 +104,21 @@ class AuthorService(
             .map { authorMapper.toDto(it) }
     }
 
+    @Transactional
     fun delete(id: String) {
         logger.info("Deleting author by id: $id")
 
         val author =  authorRepository.findById(id)
             .orElseThrow{ NotFoundException("Author not found with id: $id") }
 
-        try {
-            logger.info("Deleting author image and thumbnail")
-            Files.delete(Path.of(fileProperties.base +"/"+ author.imagePath))
-            Files.delete(Path.of(fileProperties.base +"/"+ author.thumbnailPath))
-            Files.delete(Path.of(fileProperties.base +"/"+ author.imagePath).parent)
-        } catch (e: Exception) {
-            logger.error("Error deleting file", e.message)
-        } finally {
-            logger.info("Deleting author from database")
+        if(author.deleted) {
+            Files.delete(Path.of(fileProperties.base, author.imageFile?.path))
+            Files.delete(Path.of(fileProperties.base,author.thumbFile?.path))
+            Files.delete(Path.of(fileProperties.base, author.imageFile?.path).parent)
             authorRepository.delete(author);
+        } else {
+            author.deleted = true
+            authorRepository.save(author)
         }
     }
