+++ src/main/kotlin/com/marufh/pathagar/book/service/BookService.kt
@@ -4,9 +4,13 @@ import com.marufh.pathagar.author.repository.AuthorRepository
 import com.marufh.pathagar.book.dto.BookDto
 import com.marufh.pathagar.book.dto.BookMapper
 import com.marufh.pathagar.book.repository.BookRepository
-import com.marufh.pathagar.category.CategoryRepository
+import com.marufh.pathagar.category.model.CategoryRepository
 import com.marufh.pathagar.config.FileProperties
-import jakarta.persistence.EntityNotFoundException
+import com.marufh.pathagar.exception.NotFoundException
+import com.marufh.pathagar.file.dto.FileDto
+import com.marufh.pathagar.file.entity.FileType
+import com.marufh.pathagar.file.service.FileUploadService
+import com.marufh.pathagar.file.service.PdfService
 import org.slf4j.LoggerFactory
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -17,9 +21,11 @@ import java.nio.file.Path
 
 @Service
 class BookService(
+    private val pdfService: PdfService,
+    private val bookMapper: BookMapper,
     private val bookRepository: BookRepository,
     private val fileProperties: FileProperties,
-    private val bookMapper: BookMapper,
+    private val fileUploadService: FileUploadService,
     private val categoryRepository: CategoryRepository,
     private val authorRepository: AuthorRepository) {
 
@@ -27,8 +33,24 @@ class BookService(
 
     fun create(bookDto: BookDto): BookDto {
         logger.info("Creating book: ${bookDto.name}")
-        bookMapper.toEntity(bookDto).run {
-            return bookMapper.toDto(bookRepository.save(this))
+
+        if(bookDto.file == null) {
+            throw IllegalArgumentException("File is required")
+        }
+
+        val pdfFile = fileUploadService.createFile(FileDto(
+            name = bookDto.name,
+            file = bookDto.file!!,
+            fileType = FileType.BOOK
+        ))
+        val coverImageFile = pdfService.createCoverImage(Path.of(fileProperties.base +"/"+ pdfFile.path))
+        val totalPage = pdfService.getTotalPage(Path.of(fileProperties.base +"/"+ pdfFile.path).toFile())
+
+        return bookMapper.toEntity(bookDto).let {
+            it.pdfFile = pdfFile
+            it.coverImage = coverImageFile
+            it.totalPage = totalPage
+            bookMapper.toDto(bookRepository.save(it))
         }
     }
 
@@ -37,10 +59,19 @@ class BookService(
         logger.info("Updating book: ${bookDto.name}")
 
         val book = bookRepository.findById(bookDto.id!!)
-            .orElseThrow { EntityNotFoundException("Book not found with id: ${bookDto.id}") }
+            .orElseThrow { NotFoundException("Book not found with id: ${bookDto.id}") }
 
         book.name = bookDto.name
         book.description = bookDto.description
+        if(bookDto.file != null) {
+            book.pdfFile = fileUploadService.createFile(FileDto(
+                name = bookDto.name,
+                file = bookDto.file!!,
+                fileType = FileType.BOOK
+            ))
+            book.coverImage = pdfService.createCoverImage(Path.of(fileProperties.base +"/"+ book.pdfFile?.path))
+            book.totalPage = pdfService.getTotalPage(Path.of(fileProperties.base +"/"+ book.pdfFile?.path).toFile())
+        }
         return bookRepository.save(book).run { bookMapper.toDto(this) }
     }
 
@@ -50,7 +81,7 @@ class BookService(
 
         return bookRepository.findById(id)
             .map { bookMapper.toDto(it) }
-            .orElseThrow { EntityNotFoundException("Book not found with id: $id") }
+            .orElseThrow { NotFoundException("Book not found with id: $id") }
     }
 
     @Transactional
@@ -58,10 +89,10 @@ class BookService(
         logger.info("Updating book: $bookId, author: $authorId, action: $action")
 
         val book = bookRepository.findById(bookId)
-            .orElseThrow { EntityNotFoundException("Book not found with id: $bookId") }
+            .orElseThrow { NotFoundException("Book not found with id: $bookId") }
 
         val author = authorRepository.findById(authorId)
-            .orElseThrow { EntityNotFoundException("Author not found with id: $authorId") }
+            .orElseThrow { NotFoundException("Author not found with id: $authorId") }
 
         if(action == AuthorAction.ADD.action) {
             logger.info("Adding bookId: $bookId, authorId: $authorId, action: $action")
@@ -80,10 +111,10 @@ class BookService(
         logger.info("Updating book: $bookId, category: $categoryId, action: $action")
 
         val book = bookRepository.findById(bookId)
-            .orElseThrow { EntityNotFoundException("Book not found with id: $bookId") }
+            .orElseThrow { NotFoundException("Book not found with id: $bookId") }
 
         val category = categoryRepository.findById(categoryId)
-            .orElseThrow { EntityNotFoundException("Author not found with id: $categoryId") }
+            .orElseThrow { NotFoundException("Category not found with id: $categoryId") }
 
         if(action == AuthorAction.ADD.action) {
             logger.info("Adding bookId: $bookId, categoryId: $categoryId, action: $action")
@@ -109,16 +140,18 @@ class BookService(
         logger.info("Deleting book: $id")
 
        val book =  bookRepository.findById(id)
-            .orElseThrow{ EntityNotFoundException("Book not found with id: $id") }
+            .orElseThrow{ NotFoundException("Book not found with id: $id") }
 
-        try {
-            Files.delete(Path.of(fileProperties.base +"/"+ book.filePath))
-            Files.delete(Path.of(fileProperties.base +"/"+ book.coverImage))
-            Files.delete(Path.of(fileProperties.base +"/"+ book.filePath).parent)
-        } catch (e: Exception) {
-            logger.info("Error deleting file: ${e.message}")
-        } finally {
+        if(book.deleted == true) {
+            // Hard delete
+            Files.delete(Path.of(fileProperties.base, book.pdfFile?.path))
+            Files.delete(Path.of(fileProperties.base, book.coverImage?.path))
+            Files.delete(Path.of(fileProperties.base, book.pdfFile?.path).parent)
             bookRepository.delete(book);
+        } else {
+            // Soft delete
+            book.deleted = true
+            bookRepository.save(book)
         }
     }
